+++ R/ODPSCP-package.R
@@ -1,0 +1,14 @@
+#' @title ODPSCP
+#'
+#' @description
+#' The ODPSCP shiny interface comes in a series of UI and server plus helper functions
+#' that support the standardized reporting of prioritization results.
+#'
+#' @name ODPSCP
+#' @docType package
+#' @keywords internal
+"_PACKAGE"
+
+globalVariables(c(
+  "element"
+))