+++ NEWS.md
@@ -3,6 +3,8 @@
 * Incorporated Expert-Feedback and updated the protocol with additional fields and explanations.
 * Added glossary, to be further expanded in future versions #4
 * Added tooltip to the protocol for each fields (can be toggled off) #6
+* Added option for enabling study region delineation #2
+* Added option to export a word or pdf summary of the protocol #5
 * Added a bookmark option to store the current setting of the protocol as server id (experimental)
 
 ## Version 0.2